requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv>=1.0.1,<2.0.0
supabase==1.0.3
openai==1.3.0
pydantic>=2.8,!=2.10
python-multipart==0.0.6
httpx==0.23.3
llama-cloud-services==0.6.51
streamlit==1.28.0
streamlit-chat==0.1.1
requests==2.31.0
